move gui into its own file
